fix:working on chormatic
