+++ .storybook/main.ts
@@ -6,15 +6,17 @@ const config: StorybookConfig = {
     "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"
   ],
   "addons": [
-    "@storybook/addon-essentials",
     "@storybook/addon-onboarding",
     "@chromatic-com/storybook",
-    "@storybook/experimental-addon-test"
+    "@storybook/addon-docs",
+    "@storybook/addon-a11y",
+    "@storybook/addon-vitest",
+    '@storybook/addon-themes',
+    '@storybook-addon-performance'
   ],
   "framework": {
     "name": "@storybook/react-vite",
     "options": {}
-  },
- 
+  }
 };
 export default config;
