--- .storybook/mode.ts
@@ -1,20 +1,0 @@
-export const allModes = {
-  mobile_small: {
-    name: "Mobile S",
-    styles: { width: "375px", height: "667px" },
-  },
-  mobile_medium: {
-    name: "Mobile M",
-    styles: { width: "425px", height: "736px" },
-  },
-  tablet: { name: "Tablet", styles: { width: "768px", height: "1024px" } },
-  laptop_small: {
-    name: "Laptop S",
-    styles: { width: "1024px", height: "768px" },
-  },
-  default: { name: "Default", styles: { width: "1280px", height: "720px" } },
-  laptop_large: {
-    name: "Laptop L",
-    styles: { width: "1440px", height: "900px" },
-  },
-};
