+++ src/App.tsx
@@ -2,7 +2,6 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import { Button } from './components/Button'
 
 function App() {
   const [count, setCount] = useState(0)
@@ -10,7 +9,6 @@ function App() {
   return (
     <>
       <div>
-       <Button label ={'Submit'}/>
         <a href="https://vite.dev" target="_blank">
           <img src={viteLogo} className="logo" alt="Vite logo" />
         </a>
