+++ src/components/Button.stories.ts
@@ -1,23 +1,28 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-vite';
 import { Button } from './Button';
 
 const meta = {
-    title : 'Button',
-    component : Button,
-    tags: ['autodocs'],
-} satisfies Meta<typeof Button>
+  title: 'Example/Button',
+  component: Button,
+  parameters: {
+
+    layout: 'centered',
+  },
+
+} satisfies Meta<typeof Button>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Primary : Story ={
-    args : {
-        label :"Primary"
-    }
-}
+export const Primary: Story = {
+  args: {
+    label: 'Button',
+  },
+};
 
-export const Secondary : Story ={
-    args : {
-        label :"Secondary"
-    }
-}
+export const Secondary: Story = {
+  args: {
+    label: 'Button',
+  },
+};
+
