--- tailwind.config.js
@@ -1,8 +1,0 @@
- /** @type {import('tailwindcss').Config} */
- export default {
-  content: ["./src/**/*.{html,js,tsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
