+++ vitest.shims.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="@vitest/browser/providers/playwright" />
