+++ vitest.workspace.ts
@@ -3,19 +3,19 @@ import { fileURLToPath } from 'node:url';
 
 import { defineWorkspace } from 'vitest/config';
 
-import { storybookTest } from '@storybook/experimental-addon-test/vitest-plugin';
+import { storybookTest } from '@storybook/addon-vitest/vitest-plugin';
 
 const dirname =
   typeof __dirname !== 'undefined' ? __dirname : path.dirname(fileURLToPath(import.meta.url));
 
-// More info at: https://storybook.js.org/docs/writing-tests/test-addon
+// More info at: https://storybook.js.org/docs/next/writing-tests/integrations/vitest-addon
 export default defineWorkspace([
   'vite.config.ts',
   {
     extends: 'vite.config.ts',
     plugins: [
       // The plugin will run tests for the stories defined in your Storybook config
-      // See options at: https://storybook.js.org/docs/writing-tests/test-addon#storybooktest
+      // See options at: https://storybook.js.org/docs/next/writing-tests/integrations/vitest-addon#storybooktest
       storybookTest({ configDir: path.join(dirname, '.storybook') }),
     ],
     test: {
@@ -23,8 +23,8 @@ export default defineWorkspace([
       browser: {
         enabled: true,
         headless: true,
-        name: 'chromium',
-        provider: 'playwright'
+        provider: 'playwright',
+        instances: [{ browser: 'chromium' }]
       },
       setupFiles: ['.storybook/vitest.setup.ts'],
     },
